requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pytz==2021.1
rpi==0.0.1
RPi.GPIO==0.7.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
